+++ components/archive/impl/src/main/java/com/flipperdevices/archive/impl/model/ArchiveNavigationConfig.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.archive.impl.model
 
 import com.flipperdevices.archive.model.CategoryType
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import com.flipperdevices.deeplink.model.Deeplink
 import kotlinx.serialization.Serializable
 
@@ -14,6 +15,9 @@ sealed class ArchiveNavigationConfig {
         val categoryType: CategoryType,
         val deeplink: Deeplink.BottomBar.ArchiveTab.ArchiveCategory?
     ) : ArchiveNavigationConfig()
+
+    @Serializable
+    data class InArchiveRemoteControl(val keyPath: FlipperKeyPath) : ArchiveNavigationConfig()
 
     @Serializable
     data object OpenSearch : ArchiveNavigationConfig()
@@ -34,8 +38,16 @@ fun Deeplink.BottomBar.ArchiveTab?.toArchiveNavigationStack(): List<ArchiveNavig
             }
         }
 
-        null -> {
+        is Deeplink.BottomBar.ArchiveTab.ArchiveCategory.OpenSavedRemoteControl -> {
+            val fileType = category
+            if (fileType != null) {
+                stack.add(
+                    ArchiveNavigationConfig.InArchiveRemoteControl(keyPath)
+                )
+            }
         }
+
+        null -> {}
     }
     return stack
 }