+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/api/BottomBarDecomposeComponentImpl.kt
@@ -37,6 +37,7 @@ import com.flipperdevices.notification.api.FlipperAppNotificationDialogApi
 import com.flipperdevices.toolstab.api.ToolsDecomposeComponent
 import com.flipperdevices.ui.decompose.DecomposeComponent
 import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
+import com.flipperdevices.ui.decompose.findChildByConfig
 import com.flipperdevices.ui.decompose.findComponentByConfig
 import com.flipperdevices.ui.decompose.popOr
 import com.flipperdevices.unhandledexception.api.UnhandledExceptionRenderApi
@@ -137,7 +138,8 @@ class BottomBarDecomposeComponentImpl @AssistedInject constructor(
         is BottomBarTabConfig.Tools -> toolsScreenFactory(
             componentContext = componentContext,
             deeplink = config.deeplink,
-            onBack = { navigation.popOr(onBack::invoke) }
+            onBack = { navigation.popOr(onBack::invoke) },
+            onDeeplink = ::handleDeeplink
         )
 
         is BottomBarTabConfig.Apps -> fapHubScreenFactory(
@@ -174,10 +176,12 @@ class BottomBarDecomposeComponentImpl @AssistedInject constructor(
     override fun handleDeeplink(deeplink: Deeplink.BottomBar) {
         when (deeplink) {
             is Deeplink.BottomBar.ArchiveTab -> {
-                val instance = stack.findComponentByConfig(BottomBarTabConfig.Archive::class)
+                val child = stack.findChildByConfig(BottomBarTabConfig.Archive::class)
+                val instance = child?.instance
                 if (instance == null || instance !is ArchiveDecomposeComponent<*>) {
                     navigation.bringToFront(BottomBarTabConfig.Archive(deeplink))
                 } else {
+                    navigation.bringToFront(child.configuration)
                     instance.handleDeeplink(deeplink)
                 }
             }