+++ components/remote-controls/grid/remote/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/presentation/decompose/internal/RemoteGridComponentImpl.kt
@@ -10,7 +10,7 @@ import com.flipperdevices.keyedit.api.NotSavedFlipperFile
 import com.flipperdevices.keyedit.api.NotSavedFlipperKey
 import com.flipperdevices.remotecontrols.api.DispatchSignalApi
 import com.flipperdevices.remotecontrols.api.SaveTempSignalApi
-import com.flipperdevices.remotecontrols.api.model.ServerRemoteControlParam
+import com.flipperdevices.remotecontrols.grid.remote.api.model.ServerRemoteControlParam
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.decompose.RemoteGridComponent
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.mapping.GridComponentStateMapper
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.viewmodel.ConnectionViewModel
@@ -101,6 +101,8 @@ class RemoteGridComponentImpl @AssistedInject constructor(
         dispatchSignalApi.dismissBusyDialog()
     }
 
+    private fun FlipperFilePath.toNonTempPath() = copy(folder = folder.replace("/temp", ""))
+
     override fun save() {
         val rawRemotes = remoteGridViewModel.getRawRemotesContent() ?: return
         val rawUi = remoteGridViewModel.getRawPagesContent() ?: return
@@ -110,7 +112,7 @@ class RemoteGridComponentImpl @AssistedInject constructor(
                 FlipperFilePath(
                     folder = param.extTempFolderPath,
                     nameWithExtension = param.nameWithExtension
-                ),
+                ).toNonTempPath(),
                 content = FlipperKeyContent.RawData(rawRemotes.toByteArray())
             ),
             additionalFiles = listOf(
@@ -118,7 +120,7 @@ class RemoteGridComponentImpl @AssistedInject constructor(
                     path = FlipperFilePath(
                         folder = param.extTempFolderPath,
                         nameWithExtension = param.uiFileNameWithExtension
-                    ),
+                    ).toNonTempPath(),
                     content = FlipperKeyContent.RawData(rawUi.toByteArray())
                 )
             ),
@@ -127,18 +129,30 @@ class RemoteGridComponentImpl @AssistedInject constructor(
         onSaveKey.invoke(notSavedFlipperFile)
     }
 
-    override fun onButtonClick(identifier: IfrKeyIdentifier) {
-        val gridLoadedState =
-            (remoteGridViewModel.state.value as? RemoteGridViewModel.State.Loaded) ?: return
-        val remotes = gridLoadedState.remotes
+    private fun onButtonClick(identifier: IfrKeyIdentifier, isOneTime: Boolean) {
+        val gridLoadedState = (remoteGridViewModel.state.value as? RemoteGridViewModel.State.Loaded)
+            ?: return
         dispatchSignalApi.dispatch(
             identifier = identifier,
-            remotes = remotes,
+            remotes = gridLoadedState.remotes,
+            isOneTime = isOneTime,
             ffPath = FlipperFilePath(
                 folder = param.extTempFolderPath,
                 nameWithExtension = param.nameWithExtension
             )
         )
+    }
+
+    override fun onButtonClick(identifier: IfrKeyIdentifier) {
+        onButtonClick(identifier, true)
+    }
+
+    override fun onButtonLongClick(identifier: IfrKeyIdentifier) {
+        onButtonClick(identifier, false)
+    }
+
+    override fun onButtonRelease() {
+        dispatchSignalApi.stopEmulate()
     }
 
     override fun tryLoad() = remoteGridViewModel.tryLoad()