+++ components/remote-controls/grid/saved/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/local/presentation/decompose/internal/LocalGridComponentImpl.kt
@@ -64,19 +64,32 @@ class LocalGridComponentImpl @AssistedInject constructor(
         }
     ).stateIn(coroutineScope, SharingStarted.Eagerly, LocalGridComponent.Model.Loading)
 
-    override fun onButtonClick(identifier: IfrKeyIdentifier) {
-        val gridLoadedState =
-            (localGridViewModel.state.value as? LocalGridViewModel.State.Loaded) ?: return
-        val remotes = gridLoadedState.remotes
+    private fun onButtonClick(identifier: IfrKeyIdentifier, isOneTime: Boolean) {
+        val gridLoadedState = (localGridViewModel.state.value as? LocalGridViewModel.State.Loaded)
+            ?: return
 
         dispatchSignalApi.dispatch(
             identifier = identifier,
-            remotes = remotes,
-            ffPath = gridLoadedState.keyPath.path
+            remotes = gridLoadedState.remotes,
+            ffPath = gridLoadedState.keyPath.path,
+            isOneTime = isOneTime
         )
     }
 
-    override fun onRename(onEndAction: (FlipperKeyPath) -> Unit) = localGridViewModel.onRename(onEndAction)
+    override fun onButtonClick(identifier: IfrKeyIdentifier) {
+        onButtonClick(identifier, true)
+    }
+
+    override fun onButtonLongClick(identifier: IfrKeyIdentifier) {
+        onButtonClick(identifier, false)
+    }
+
+    override fun onButtonRelease() {
+        dispatchSignalApi.stopEmulate()
+    }
+
+    override fun onRename(onEndAction: (FlipperKeyPath) -> Unit) =
+        localGridViewModel.onRename(onEndAction)
 
     override fun onDelete(onEndAction: () -> Unit) = localGridViewModel.onDelete(onEndAction)