+++ components/toolstab/impl/src/main/java/com/flipperdevices/toolstab/impl/api/ToolsMainScreenDecomposeComponentImpl.kt
@@ -7,9 +7,10 @@ import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.router.stack.StackNavigation
 import com.arkivanov.decompose.router.stack.pushToFront
 import com.flipperdevices.core.ui.lifecycle.viewModelWithFactory
+import com.flipperdevices.metric.api.MetricApi
+import com.flipperdevices.metric.api.events.SimpleEvent
 import com.flipperdevices.toolstab.impl.composable.ComposableHub
 import com.flipperdevices.toolstab.impl.model.ToolsNavigationConfig
-import com.flipperdevices.toolstab.impl.viewmodel.DebugSettingsViewModel
 import com.flipperdevices.toolstab.impl.viewmodel.ToolsNotificationViewModel
 import com.flipperdevices.ui.decompose.ScreenDecomposeComponent
 import dagger.assisted.Assisted
@@ -21,7 +22,7 @@ class ToolsMainScreenDecomposeComponentImpl @AssistedInject constructor(
     @Assisted componentContext: ComponentContext,
     @Assisted private val navigation: StackNavigation<ToolsNavigationConfig>,
     private val toolsNotificationViewModelProvider: Provider<ToolsNotificationViewModel>,
-    private val debugSettingsViewModelProvider: Provider<DebugSettingsViewModel>,
+    private val metricApi: MetricApi
 ) : ScreenDecomposeComponent(componentContext) {
     @Composable
     @Suppress("NonSkippableComposable")
@@ -29,19 +30,15 @@ class ToolsMainScreenDecomposeComponentImpl @AssistedInject constructor(
         val nfcAttackViewModel = viewModelWithFactory(key = null) {
             toolsNotificationViewModelProvider.get()
         }
-        val debugSettingsViewModel = viewModelWithFactory(key = null) {
-            debugSettingsViewModelProvider.get()
-        }
         val hasNotification by nfcAttackViewModel.hasNotificationStateFlow.collectAsState()
-        val showRemoteControls by debugSettingsViewModel.showRemoteControls.collectAsState()
 
         ComposableHub(
             hasNotification = hasNotification,
-            showRemoteControls = showRemoteControls,
             onOpenMfKey32 = {
                 navigation.pushToFront(ToolsNavigationConfig.MfKey32)
             },
             onOpenRemoteControls = {
+                metricApi.reportSimpleEvent(SimpleEvent.OPEN_INFRARED_LIBRARY)
                 navigation.pushToFront(ToolsNavigationConfig.RemoteControls)
             }
         )